+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 #### API
 
+- Add `DDS-CNF` density method to `DensityArchive` ({pr}`691`)
 - Add PyTorch to ribs[all] deps ({pr}`692`)
 - Add `NSLCRanker` for Novelty Search with Local Competition ({pr}`690`)